Move session components to new folder
